Python 3 port: implicit conversion of int from float warning
Python 3 is generating this deprecation warning about floats being
passed to a pygame API which is expecting integers:
    AsteroidsInfinity.py:276: DeprecationWarning: an integer is required (got type float).  Implicit conversion to integers using __int__ is deprecated, and may be removed in a future version of Python.
      pygame.draw.polygon(surface, backcolor, temp_points)

When running AsteroidsInfinity under Python 2 floats were passed too, no
warning was raised.  Explicitly convert floats to ints.

+++ AsteroidsInfinity.py
@@ -271,7 +271,8 @@ class Obj(pygame.sprite.Sprite):
                 points_list = points_list[1:]
             temp_points = []
             for point in points_list:
-                temp_points.append((math.sin(point[0] + self.angle) * point[1] + render_pos[0], math.cos(point[0] + self.angle) * point[1] + render_pos[1]))
+                temp_points.append((int(math.sin(point[0] + self.angle) * point[1] + render_pos[0]),
+                                    int(math.cos(point[0] + self.angle) * point[1] + render_pos[1])))
             pygame.draw.polygon(surface, backcolor, temp_points)
             dirtyrects.append(pygame.draw.aalines(surface, color, 1, temp_points).inflate(2, 2))
             #pygame.draw.circle(surface, (100, 100, 100), render_pos, self.radius, 1)
